+++ h2mare/processing/__init__.py
@@ -6,7 +6,7 @@ from .core.cds import (
     process_radiation,
     process_waves,
 )
-from .core.cmems import process_chl, process_mld, process_ssh, process_sst
+from .core.cmems import process_chl, process_ssh, process_sst
 from .core.fronts import FrontProcessor, apply_boa_fronts
 from .extractor import Extractor, ensure_row_id
 
@@ -19,7 +19,6 @@ __all__ = [
     "apply_boa_fronts",
     "process_fsle",
     "process_chl",
-    "process_mld",
     "process_ssh",
     "process_sst",
     "process_atm_accum_avg",